extract the funtion cleanUrlParam from app.tsx
