+++ src/App.tsx
@@ -5,7 +5,7 @@ import { FilesProvider } from './contexts/Files';
 import DownloadView from './views/DownloadView';
 import HomeView from './views/HomeView';
 import NotFoundView from './views/NotFoundView';
-import { useEffect } from 'react';
+import { cleanUrlParam } from './utils/cleanUrl';
 
 function DownloadRedirectWrapper() {
   const { sendId } = useParams();
@@ -16,27 +16,7 @@ function DownloadRedirectWrapper() {
 }
 
 function App() {
-  useEffect(() => {
-    const url = new URL(window.location.href);
-    let hadChanges = false;
-
-    const currentKeys = Array.from(url.searchParams.keys());
-
-    currentKeys.forEach((param) => {
-      if (param === '_gl' || param === '_gcl_au' || param === '_fplc' || param.startsWith('_ga')) {
-        url.searchParams.delete(param);
-        hadChanges = true;
-      }
-    });
-
-    if (hadChanges) {
-      const newUrl = url.searchParams.toString()
-        ? `${url.pathname}?${url.searchParams.toString()}${url.hash}`
-        : `${url.pathname}${url.hash}`;
-
-      window.history.replaceState({}, document.title, newUrl);
-    }
-  }, []);
+  cleanUrlParam();
 
   return (
     <>
